+++ 20250110/pinia/package.json
@@ -9,6 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "pinia": "^2.3.0",
     "vue": "^3.5.13"
   },
   "devDependencies": {
