+++ 20250110/pinia/src/main.js
@@ -1,6 +1,10 @@
 import "./assets/css/index.css";
 
+import { createPinia } from "pinia";
 import { createApp } from "vue";
 import App from "./App.vue";
 
-createApp(App).mount("#app");
+const pinia = createPinia();
+const app = createApp(App);
+app.use(pinia);
+app.mount("#app");
